requirements.txt
python-dotenv
langchain
pinecone-client
fastapi
langchain-together
langchain-pinecone
uvicorn
langchain-community
pycryptodome